+++ changes.md
@@ -13,6 +13,7 @@
 1. Upgrade three.js r174 -> r180
 1. Upgrade lil-gui.mjs 0.19.2 -> 0.20.0
 1. Use ES6 modules to implement geoworker, enable node.js usage
+1. Remove countGeometryFaces function - use numGeometryFaces instead
 1. Remove experimental RHist classes, deprecated in ROOT 6.38
 1. Internal - ws members are private, new methods has to be used
 1. Fix - ticks size and labels with kMoreLogLabels axis bit